Fixed code that assumed typedef unsigned long uint64_t;
On OS X, it's typedef'd to unsigned long long instead.

+++ dataCollect/InputParser.cpp
@@ -261,7 +261,7 @@ void bindDIFF(parserData *d, field_info *fi, PacketDiff &val) {
     case FT_UINT48:
     case FT_UINT56:
     case FT_UINT64:
-      val = PacketDiff(i, si, fi->value.value.uinteger64);
+      val = PacketDiff(i, si, (uint64_t)fi->value.value.uinteger64);
       DPRINT(d, fi, debugStr + std::to_string(fi->value.value.uinteger64), "");
       return;
